requirements.txt
tensorflow-gpu==2.3.0
tqdm==4.50.2
pandas==1.1.3
numpy==1.19.3
scikit-learn==0.23.2